# api/openapi.yaml
openapi: 3.0.3
info:
  title: NFL Team Efficiency Dashboard API
  version: "1.0.0"
  description: >
    REST API serving team efficiency metrics (EPA/play, success rate,
    points/drive — offense, defense, and special teams), head-to-head
    comparisons, leaderboards, and game win-probability predictions,
    computed from nflverse play-by-play data.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/teams/{abbreviation}/stats:
    get:
      summary: Weekly efficiency trend for one team
      parameters:
        - name: abbreviation
          in: path
          required: true
          schema:
            type: string
          example: KC
        - $ref: "#/components/parameters/Season"
      responses:
        "200":
          description: Team info plus one row per week played
          content:
            application/json:
              schema:
                type: object
                required: [team, season, weeks]
                properties:
                  team:
                    $ref: "#/components/schemas/Team"
                  season:
                    type: integer
                  weeks:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeamWeekMetrics"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Season:
      name: season
      in: query
      required: false
      schema:
        type: integer
      description: Defaults to the most recent season with data
  responses:
    NotFound:
      description: No team with that abbreviation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Team:
      type: object
      required: [id, name, abbreviation, conference, division, logo_url, color, color2]
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        conference:
          type: string
          nullable: true
        division:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
          description: Primary team color, hex
        color2:
          type: string
          nullable: true
          description: Secondary team color, hex
...
    TeamWeekMetrics:
      type: object
      required:
        - week
        - epa_per_play
        - success_rate
        - points_per_drive
        - epa_per_play_allowed
        - success_rate_allowed
        - points_per_drive_allowed
        - st_epa_per_play
        - st_epa_per_play_allowed
      properties:
        week:
          type: integer
        epa_per_play:
          type: number
          nullable: true
        success_rate:
          type: number
          nullable: true
        points_per_drive:
          type: number
          nullable: true
        epa_per_play_allowed:
          type: number
          nullable: true
        success_rate_allowed:
          type: number
          nullable: true
        points_per_drive_allowed:
          type: number
          nullable: true
        st_epa_per_play:
          type: number
          nullable: true
        st_epa_per_play_allowed:
          type: number
          nullable: true
...